requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
colorama==0.4.6
exceptiongroup==1.2.2
fsrs==5.1.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
ollama==0.4.7
packaging==24.2
pluggy==1.5.0
pydantic==2.10.6
pydantic-core==2.27.2
pytest==8.3.4
PyYAML==6.0.2
sniffio==1.3.1
tomli==2.2.1
typing-extensions==4.12.2
